Local test refactor, fixes #371, fixes #381 (#383)
* move local start/shutdown to TestMain so tests can be run individually

* fixes for test interdependencies

* remove should not fail if there is no data to be removed

* only fail stop if no containers are found

* use env to set short mode, test fixes and cleanup

+++ pkg/plugins/platform/local.go
@@ -629,8 +629,8 @@ func (l *LocalApp) DockerEnv() {
 func (l *LocalApp) Stop() error {
 	l.DockerEnv()
 
-	if l.SiteStatus() != SiteRunning {
-		return fmt.Errorf("site does not appear to be running - web container %s", l.SiteStatus())
+	if l.SiteStatus() == SiteNotFound {
+		return fmt.Errorf("no site to remove")
 	}
 
 	err := dockerutil.ComposeCmd(l.ComposeFiles(), "stop")
@@ -745,16 +745,20 @@ func (l *LocalApp) Down(removeData bool) error {
 		dir := filepath.Dir(l.AppConfig.DataDir)
 		// mysql data can be set to read-only on linux hosts. PurgeDirectory ensures files
 		// are writable before we attempt to remove them.
-		err := fileutil.PurgeDirectory(dir)
-		if err != nil {
-			return fmt.Errorf("failed to remove data directories: %v", err)
+		if !fileutil.FileExists(dir) {
+			util.Warning("No application data to remove")
+		} else {
+			err := fileutil.PurgeDirectory(dir)
+			if err != nil {
+				return fmt.Errorf("failed to remove data directories: %v", err)
+			}
+			// PurgeDirectory leaves the directory itself in place, so we remove it here.
+			err = os.RemoveAll(dir)
+			if err != nil {
+				return fmt.Errorf("failed to remove data directories: %v", err)
+			}
+			util.Success("Application data removed")
 		}
-		// PurgeDirectory leaves the directory itself in place, so we remove it here.
-		err = os.RemoveAll(dir)
-		if err != nil {
-			return fmt.Errorf("failed to remove data directories: %v", err)
-		}
-		util.Success("Application data removed")
 	}
 
 	err := dockerutil.ComposeCmd(l.ComposeFiles(), "down", "-v")
